99.cuopt/requirements.txt
# cuOpt requirements for local installation
cuopt-cu11>=23.12.0
cudf-cu11>=23.12.0
pandas>=1.3.0
numpy>=1.21.0

# Database ORM
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
alembic>=1.13.0

# Prisma ORM
prisma>=0.11.0
python-dotenv>=1.0.0

# Alternative for CUDA 12
# cuopt-cu12>=23.12.0
# cudf-cu12>=23.12.0